# frontend_admin/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Awesome Template API","description":"Awesome Template バックエンドサービスの API ドキュメントです。","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Admin認証","description":"Admin認証に関連する操作です"},{"name":"E2Eテスト支援","description":"E2Eテストで利用するデータセットアップの補助APIです"},{"name":"認証","description":"ユーザー認証・登録に関連する操作です"},{"name":"ユーザー","description":"ユーザー管理に関連する操作です"},{"name":"ユーザー","description":"ユーザーを管理するための操作です"}],"paths":{"/v1/customer/users/changePassword":{"put":{"tags":["ユーザー"],"summary":"パスワード変更","description":"認証済みユーザーのパスワードを変更します","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"パスワードの変更に成功しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordResponse"}}}},"400":{"description":"リクエストが不正です"},"401":{"description":"認証が必要です"}}}},"/v1/customer/users/changeEmail":{"put":{"tags":["ユーザー"],"summary":"メールアドレス変更","description":"認証済みユーザーのメールアドレスを変更します","operationId":"changeEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailRequest"}}},"required":true},"responses":{"200":{"description":"メールアドレスの変更に成功しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailResponse"}}}},"400":{"description":"リクエストが不正です"},"401":{"description":"認証が必要です"}}}},"/v1/customer/auth/signup":{"post":{"tags":["認証"],"summary":"ユーザー登録","description":"指定した情報で新しいユーザーを登録します。","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"ユーザーの登録に成功しました。","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}},"400":{"description":"リクエストペイロードが不正です。"}}}},"/v1/customer/auth/signin":{"post":{"tags":["認証"],"summary":"サインイン","description":"メールアドレスとパスワードでサインインします。","operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninRequest"}}},"required":true},"responses":{"200":{"description":"サインインに成功しました。","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponse"}}}},"400":{"description":"メールアドレスまたはパスワードが正しくありません。"}}}},"/v1/admin/auth/signin":{"post":{"tags":["Admin認証"],"summary":"Adminサインイン","description":"Admin管理者としてサインインします","operationId":"adminSignin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminSigninRequest"}}},"required":true},"responses":{"200":{"description":"サインインに成功しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminSigninResponse"}}}},"400":{"description":"リクエストが不正です"},"401":{"description":"認証に失敗しました"}}}},"/e2e/trigger_error":{"post":{"tags":["E2Eテスト支援"],"summary":"500エラーを発生","description":"500エラーログのテスト用。意図的に例外を発生させます。","operationId":"triggerError","responses":{"200":{"description":"OK"}}}},"/e2e/reset_data":{"post":{"tags":["E2Eテスト支援"],"summary":"DBをリセット","description":"アプリケーションが管理するテーブルからデータをすべて削除します。","operationId":"resetData","responses":{"200":{"description":"OK"}}}},"/e2e/dummy_token":{"post":{"tags":["E2Eテスト支援"],"summary":"E2Eテスト用トークン生成","description":"E2Eテストでauth guardをバイパスするためのダミートークンを生成します。setCookie=trueの場合、httpOnly Cookieとしても設定します。","operationId":"generateDummyToken","parameters":[{"name":"setCookie","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TestTokenResponse"}}}}}}},"/e2e/create_data":{"post":{"tags":["E2Eテスト支援"],"summary":"テストデータをセットアップ","description":"指定されたテーブルにテストデータを1レコード作成します。","operationId":"createData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetupDataRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/customer/users":{"get":{"tags":["ユーザー"],"summary":"ユーザー一覧を取得","description":"登録されているすべてのユーザーを取得します。","operationId":"getAllUsers","responses":{"200":{"description":"ユーザーの取得に成功しました。","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindAllUsersResponse"}}}}}}},"/v1/customer/auth/me":{"get":{"tags":["認証"],"summary":"認証済みユーザー情報取得","description":"現在認証されているユーザーの情報を取得します。","operationId":"me","responses":{"200":{"description":"ユーザー情報の取得に成功しました。","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}},"401":{"description":"認証されていません。"}}}},"/v1/admin/auth/me":{"get":{"tags":["Admin認証"],"summary":"現在のAdmin情報取得","description":"認証済みAdminの情報を取得します","operationId":"adminMe","responses":{"200":{"description":"Admin情報の取得に成功しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminFindMeResponse"}}}},"401":{"description":"認証が必要です"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"ChangePasswordRequest":{"type":"object","description":"パスワード変更リクエスト","properties":{"currentPassword":{"type":"string","description":"現在のパスワード","example":"currentPassword123","minLength":1},"newPassword":{"type":"string","description":"新しいパスワード（8文字以上）","example":"newPassword456","minLength":1}},"required":["currentPassword","newPassword"]},"ChangePasswordResponse":{"type":"object","description":"パスワード変更レスポンス","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーID"},"email":{"type":"string","description":"メールアドレス","example":"user@example.com"},"updatedAt":{"type":"string","format":"date-time","description":"更新日時"}}},"ChangeEmailRequest":{"type":"object","description":"メールアドレス変更リクエスト","properties":{"email":{"type":"string","format":"email","description":"新しいメールアドレス","example":"new.email@example.com","minLength":1}},"required":["email"]},"ChangeEmailResponse":{"type":"object","description":"メールアドレス変更レスポンス","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーID"},"email":{"type":"string","description":"メールアドレス","example":"new.email@example.com"},"updatedAt":{"type":"string","format":"date-time","description":"更新日時"}}},"SignupRequest":{"type":"object","description":"新しいユーザーを登録するためのリクエストペイロードです","properties":{"email":{"type":"string","format":"email","description":"ユーザーのメールアドレス","example":"jane.doe@example.com","minLength":1},"password":{"type":"string","description":"ユーザーのパスワード","example":"SecurePassword123","minLength":1}},"required":["email","password"]},"SignupResponse":{"type":"object","description":"新しく作成されたユーザーを表します","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーの一意な識別子"},"email":{"type":"string","description":"ユーザーのメールアドレス","example":"jane.doe@example.com"},"createdAt":{"type":"string","format":"date-time","description":"ユーザーの作成日時"},"updatedAt":{"type":"string","format":"date-time","description":"ユーザーの最終更新日時"}}},"SigninRequest":{"type":"object","description":"ユーザーがサインインするためのリクエストペイロードです","properties":{"email":{"type":"string","format":"email","description":"ユーザーのメールアドレス","example":"jane.doe@example.com","minLength":1},"password":{"type":"string","description":"ユーザーのパスワード","example":"SecurePassword123","minLength":1}},"required":["email","password"]},"SigninResponse":{"type":"object","description":"サインインしたユーザーを表します","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーの一意な識別子"},"email":{"type":"string","description":"ユーザーのメールアドレス","example":"jane.doe@example.com"},"accessToken":{"type":"string","description":"JWT認証トークン","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"createdAt":{"type":"string","format":"date-time","description":"ユーザーの作成日時"},"updatedAt":{"type":"string","format":"date-time","description":"ユーザーの最終更新日時"}}},"AdminSigninRequest":{"type":"object","description":"Adminサインインリクエスト","properties":{"email":{"type":"string","format":"email","description":"メールアドレス","example":"admin@example.com","minLength":1},"password":{"type":"string","description":"パスワード","example":"adminPassword123","minLength":1}},"required":["email","password"]},"AdminSigninResponse":{"type":"object","description":"Adminサインインレスポンス","properties":{"id":{"type":"string","format":"uuid","description":"Admin ID"},"email":{"type":"string","description":"メールアドレス","example":"admin@example.com"},"accessToken":{"type":"string","description":"アクセストークン"},"createdAt":{"type":"string","format":"date-time","description":"作成日時"}}},"TestTokenResponse":{"type":"object","description":"E2Eテスト用トークンのレスポンスです","properties":{"token":{"type":"string","description":"JWT認証トークン","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}},"SetupDataRequest":{"type":"object","description":"テストデータセットアップのリクエストです","properties":{"table":{"type":"string","description":"セットアップ対象のテーブル名","example":"users"}}},"UserListItem":{"type":"object","description":"一覧レスポンス内のユーザー情報です","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーの一意な識別子"},"email":{"type":"string","description":"ユーザーのメールアドレス","example":"jane.doe@example.com"},"createdAt":{"type":"string","format":"date-time","description":"ユーザーの作成日時"},"updatedAt":{"type":"string","format":"date-time","description":"ユーザーの最終更新日時"}},"required":["createdAt","email","id","updatedAt"]},"FindAllUsersResponse":{"type":"object","description":"登録済みユーザーの一覧を表します","properties":{"users":{"type":"array","description":"ユーザーの一覧です","items":{"$ref":"#/components/schemas/UserListItem","description":"ユーザーの一覧です"}}}},"MeResponse":{"type":"object","description":"認証済みユーザーの情報を表します","properties":{"id":{"type":"string","format":"uuid","description":"ユーザーの一意な識別子"},"email":{"type":"string","description":"ユーザーのメールアドレス","example":"jane.doe@example.com"}}},"AdminFindMeResponse":{"type":"object","description":"現在のAdmin情報レスポンス","properties":{"id":{"type":"string","format":"uuid","description":"Admin ID"},"email":{"type":"string","description":"メールアドレス","example":"admin@example.com"}}}}}}