# frontend_native/features/shared/api/generated/model/adminFindMeResponse.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * 現在のAdmin情報レスポンス
 */
export interface AdminFindMeResponse {
  /** メールアドレス */
  email?: string;
  /** Admin ID */
  id?: string;
}