# frontend_native/features/shared/api/generated/model/changeEmailRequest.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * メールアドレス変更リクエスト
 */
export interface ChangeEmailRequest {
  /**
   * 新しいメールアドレス
   * @minLength 1
   */
  email: string;
}