# frontend_native/features/shared/api/generated/model/changePasswordRequest.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * パスワード変更リクエスト
 */
export interface ChangePasswordRequest {
  /**
   * 現在のパスワード
   * @minLength 1
   */
  currentPassword: string;
  /**
   * 新しいパスワード（8文字以上）
   * @minLength 1
   */
  newPassword: string;
}