# frontend_web/shared/api/generated/model/adminSigninResponse.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * Adminサインインレスポンス
 */
export interface AdminSigninResponse {
  /** アクセストークン */
  accessToken?: string;
  /** 作成日時 */
  createdAt?: string;
  /** メールアドレス */
  email?: string;
  /** Admin ID */
  id?: string;
}