# frontend_web/shared/api/generated/model/changeEmailResponse.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * メールアドレス変更レスポンス
 */
export interface ChangeEmailResponse {
  /** メールアドレス */
  email?: string;
  /** ユーザーID */
  id?: string;
  /** 更新日時 */
  updatedAt?: string;
}