# frontend_web/shared/api/generated/model/changePasswordResponse.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * パスワード変更レスポンス
 */
export interface ChangePasswordResponse {
  /** メールアドレス */
  email?: string;
  /** ユーザーID */
  id?: string;
  /** 更新日時 */
  updatedAt?: string;
}